% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-class.R
\name{as.array,bayesPO_fit-method}
\alias{as.array,bayesPO_fit-method}
\title{as.array method for the bayesPO_fit class.}
\usage{
\S4method{as.array}{bayesPO_fit}(x, ...)
}
\arguments{
\item{x}{A bayesPO_fit object.}

\item{...}{Ignored.}
}
\value{
The MCMC chains organized in a way ready for the \code{bayesplot} package.
}
\description{
as.array method for the bayesPO_fit class.
}
