% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-class.R
\name{as.array.bayesPO_fit}
\alias{as.array.bayesPO_fit}
\title{Put MCMC output in an \code{array}.}
\usage{
\method{as.array}{bayesPO_fit}(x, ...)
}
\arguments{
\item{x}{Output of the \code{\link{fit_bayesPO}} function.}

\item{...}{Ignored in this version.}
}
\value{
An \code{array} with dimensions I x C x P, where I stands for number of
iterations, C for number of chains and P for total number of parameters.
P is actually larger than the number of parameters in the model, as the
the generated sizes of the latent processes and the log-posterior are also
included.
}
\description{
Prepares the MCMC results to be used by plotting functions of the
bayesplot package.
}
\seealso{
\code{\link{fit_bayesPO}}
}
