% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-class.R
\name{as.data.frame,bayesPO_fit-method}
\alias{as.data.frame,bayesPO_fit-method}
\title{as.data.frame method for the bayesPO_fit class.}
\usage{
\S4method{as.data.frame}{bayesPO_fit}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A bayesPO_fit object.}

\item{row.names}{NULL or a character vector giving the row names for the
data frame. Missing values are not allowed.}

\item{optional}{logical. If TRUE, setting row names and converting column
names to syntactic names is optional. See help('as.data.frame') for more.
Leaving as \code{FALSE} is recommended.}

\item{...}{Ignored.}
}
\value{
The MCMC chains reformatted as a data.frame
}
\description{
as.data.frame method for the bayesPO_fit class.
}
