% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-class.R
\name{as.data.frame.bayesPO_fit}
\alias{as.data.frame.bayesPO_fit}
\title{Put MCMC output in a \code{data.frame}}
\usage{
\method{as.data.frame}{bayesPO_fit}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{Output of the \code{\link{fit_bayesPO}} function.}

\item{row.names}{NULL or a character vector giving the row names for the
data frame. Missing values are not allowed.}

\item{optional}{logical. If TRUE, setting row names and converting column
names to syntactic names is optional. See help('as.data.frame') for more.
Leaving as \code{FALSE} is recommended.}

\item{...}{Ignored in this version.}
}
\value{
A data.frame where all the MCMC results are included.
}
\description{
Create a \code{data.frame} with all the MCMC results.
}
\details{
The dimension of the output is I*C x P+2, where I stands for
number of iterations, C for number of chains and P for total number of
parameters. P is actually larger than the number of parameters in the model,
as the generated sizes of the latent processes and the log-posterior are
also included.

Two extra columns are included to indicate to which chain and to which
iteration that draw belongs. This is to facilitate the use of plotting
results via the \code{ggplot2} package if desired.

If \code{row.names} is left at \code{NULL} then row names are created as
CcIi where c is the chain and i is the iteration of that row.
}
\seealso{
\code{\link{fit_bayesPO}}
}
