% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-class.R
\name{as.matrix,bayesPO_fit-method}
\alias{as.matrix,bayesPO_fit-method}
\title{as.matrix method for the bayesPO_fit class.}
\usage{
\S4method{as.matrix}{bayesPO_fit}(x, ...)
}
\arguments{
\item{x}{A bayesPO_fit object.}

\item{...}{Ignored.}
}
\value{
The MCMC chains reformatted as a matrix.
}
\description{
as.matrix method for the bayesPO_fit class.
}
