% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-class.R
\name{as.matrix.bayesPO_fit}
\alias{as.matrix.bayesPO_fit}
\title{Put MCMC output in a \code{matrix}.}
\usage{
\method{as.matrix}{bayesPO_fit}(x, ...)
}
\arguments{
\item{x}{Output of the \code{\link{fit_bayesPO}} function.}

\item{...}{Ignored in this version.}
}
\value{
A matrix where all the MCMC results are included.
}
\description{
Create a matrix with all the MCMC results.
}
\details{
The dimension of the output is I*C x P+2, where I stands for
number of iterations, C for number of chains and P for total number of
parameters. P is actually larger than the number of parameters in the model,
as the generated sizes of the latent processes and the log-posterior are
also included.

Two extra columns are included to indicate to which chain and to which
iteration that draw belongs.
}
\seealso{
\code{\link{fit_bayesPO}}
}
