% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-class.R
\docType{class}
\name{bayesPO_fit-class}
\alias{bayesPO_fit-class}
\title{Class for the result of the MCMC procedure.}
\description{
Class for the result of the MCMC procedure.
}
\section{Slots}{

\describe{
\item{\code{fit}}{The actual fit from the model. It is of class \code{\link[coda]{mcmc.list}}, as generated from the \code{coda} package.}

\item{\code{original}}{The model used to generate the chains, an object with class \code{bayesPO_model}.}

\item{\code{backgroundSummary}}{A small summary of the original background covariates. This is to ensure that continuing the chains will use the identical background matrix.}

\item{\code{area}}{A positive number indicating the area measure of the region being studied.}

\item{\code{parnames}}{The names of the parameters. If the model used selects the covariates with column names, they are replicated here. If they are the column indexes, names are generated for identification.}

\item{\code{mcmc_setup}}{The original mcmc setup used.}
}}

