% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\docType{class}
\name{bayesPO_prior-class}
\alias{bayesPO_prior-class}
\title{Joint prior class for the bayesPO package parameters.}
\description{
Joint prior class for the bayesPO package parameters.
}
\section{Slots}{

\describe{
\item{\code{beta}}{An object of BetaDeltaPrior S4 class with the appropriate Beta prior.}

\item{\code{delta}}{An object of BetaDeltaPrior S4 class with the appropriate Delta prior.}

\item{\code{lambdaStar}}{An object of LambdaStarPrior S4 class with the appropriate LambdaStar prior.}
}}

