% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\docType{class}
\name{BetaDeltaPrior-class}
\alias{BetaDeltaPrior-class}
\title{Generic class for the beta and delta parameters.}
\description{
Generic class for the beta and delta parameters.
}
\section{Slots}{

\describe{
\item{\code{family}}{The family of distributions of the prior.}
}}

