% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-class.R
\name{$,bayesPO_initial-method}
\alias{$,bayesPO_initial-method}
\title{The '$' method for the bayesPO_initial class.}
\usage{
\S4method{$}{bayesPO_initial}(x, name)
}
\arguments{
\item{x}{The bayesPO_initial object.}

\item{name}{The requested slot.}
}
\description{
The '$' method for the bayesPO_initial class.
}
