% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{$,bayesPO_prior-method}
\alias{$,bayesPO_prior-method}
\title{The '$' method for the bayesPO_prior class.}
\usage{
\S4method{$}{bayesPO_prior}(x, name)

\S4method{$}{bayesPO_prior}(x, name)
}
\arguments{
\item{x}{The bayesPO_prior object.}

\item{name}{The requested slot.}
}
\description{
The '$' method for the bayesPO_prior class.

The '$' method for the bayesPO_prior class.
}
