% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{$,GammaPrior-method}
\alias{$,GammaPrior-method}
\title{The '$' method for the GammaPrior class.}
\usage{
\S4method{$}{GammaPrior}(x, name)
}
\arguments{
\item{x}{The GammaPrior object.}

\item{name}{The requested slot.}
}
\description{
The '$' method for the GammaPrior class.
}
