% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{$,NormalPrior-method}
\alias{$,NormalPrior-method}
\title{The '$' method for the NormalPrior class.}
\usage{
\S4method{$}{NormalPrior}(x, name)
}
\arguments{
\item{x}{The NormalPrior object.}

\item{name}{The requested slot.}
}
\description{
The '$' method for the NormalPrior class.
}
