% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\name{$<-,bayesPO_model-method}
\alias{$<-,bayesPO_model-method}
\title{The '$<-' method for the bayesPO_model class.}
\usage{
\S4method{$}{bayesPO_model}(x, name) <- value
}
\arguments{
\item{x}{The bayesPO_model object.}

\item{name}{The requested slot.}

\item{value}{New value.}
}
\description{
The '$<-' method for the bayesPO_model class.
}
