% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_importance-class.R
\name{covariates_importance-class}
\alias{covariates_importance-class}
\alias{print.covariates_importance}
\alias{plot.covariates_importance}
\alias{barplot.covariates_importance}
\title{Class for covariates importance matrices}
\usage{
\method{print}{covariates_importance}(x, component = "intensity", ...)

\method{plot}{covariates_importance}(x, component = "intensity", y, ...)

\method{barplot}{covariates_importance}(x, component = "intensity", y, ...)
}
\arguments{
\item{x}{The \code{covariates_importance} object.}

\item{component}{Either \code{"intensity"}, \code{"observability"} or
\code{"both"}.}

\item{...}{Other parameters passed to \code{\link[graphics]{barplot}}.}

\item{y}{Ignored.}

\item{height}{The \code{covariates_importance} object.}
}
\description{
Objects of this class is the output of the "covariates_importance" object
from the \code{\link{bayesPO_fit-class}}. It can be plotted which uses
the \code{\link[graphics]{graphics}} package.
}
\details{
Objects of this class have two matrices. One for the intensity importance
and one for the observability importance.
}
