% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesPO.R
\name{fit_bayesPO,bayesPO_model,matrix-method}
\alias{fit_bayesPO,bayesPO_model,matrix-method}
\title{The fit_bayesPO method for the bayesPO_model class.}
\usage{
\S4method{fit_bayesPO}{bayesPO_model,matrix}(object, background, area = 1, mcmc_setup = list(n_iter = 5000))
}
\arguments{
\item{object}{A bayesPO_model object.}

\item{background}{A matrix with the covariates values in the region background.}

\item{area}{A positive number with the region's area}

\item{mcmc_setup}{A list with the components \code{burnin}, \code{thin} and
\code{n_iter} where only the latter is mandatory.}
}
\description{
The fit_bayesPO method for the bayesPO_model class.
}
