% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\docType{class}
\name{GammaPrior-class}
\alias{GammaPrior-class}
\title{Gamma prior class for the LambdaStar parameter.}
\description{
Gamma prior class for the LambdaStar parameter.
}
\section{Slots}{

\describe{
\item{\code{mu}}{The mean vector for the prior.}
}}

