% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{GammaPrior}
\alias{GammaPrior}
\title{Create a Gamma prior object for model specification.}
\usage{
GammaPrior(shape, rate)
}
\arguments{
\item{shape}{A positive number.}

\item{rate}{A positive number.}
}
\value{
A \code{GammaPrior} object with adequate slots.
}
\description{
Constructor for \code{GammaPrior-class} objects
}
