% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-class.R
\name{initialize,bayesPO_initial-method}
\alias{initialize,bayesPO_initial-method}
\title{Initialize method for bayesPO_initial objects}
\usage{
\S4method{initialize}{bayesPO_initial}(.Object, beta, delta, lambdaStar, tag)
}
\arguments{
\item{.Object}{An empty bayesPO_initial object.}

\item{beta}{A vector with initial beta values. Must be possible to coerce to
numeric.}

\item{delta}{A vector with initial delta values. Must be possible to coerce
to numeric.}

\item{lambdaStar}{A number with initial the lambdaStar value. Must be
possible to coerce to numeric.}

\item{tag}{The tag indicating the origin of the initial value.}
}
\description{
Fills the object with the necessary slots, converting them to numeric as
needed
}
