% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\name{initialize,bayesPO_model-method}
\alias{initialize,bayesPO_model-method}
\title{Initialize method for bayesPO_model objects}
\usage{
\S4method{initialize}{bayesPO_model}(
  .Object,
  po,
  intensityLink,
  intensitySelection,
  observabilityLink,
  observabilitySelection,
  init,
  prior,
  iSelectedColumns,
  oSelectedColumns
)
}
\arguments{
\item{.Object}{An empty bayesPO_model object.}

\item{po}{The matrix containing the covariates in the observed locations.}

\item{intensityLink}{A string containing the intensity Link function.}

\item{intensitySelection}{The positions of selected columns for the intensity
covariates. Can be empty if the selection was from a character vector.}

\item{observabilityLink}{A string containing the observability Link function.}

\item{observabilitySelection}{The positions of selected columns for the
observability covariates. Can be empty if the selection was from a character
vector.}

\item{init}{A bayesPO_initial object.}

\item{prior}{A bayesPO_prior object.}

\item{iSelectedColumns}{A character vector containing the selected
intensity covariates. Can be empty if the selection was from the columns
positions.}

\item{oSelectedColumns}{A character vector containing the selected
observability covariates. Can be empty if the selection was from the columns
positions.}
}
\description{
Fills the object with the necessary slots. Provides some text informing
the loaded data.
}
