% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\docType{class}
\name{LambdaStarPrior-class}
\alias{LambdaStarPrior-class}
\title{Generic class for the LambdaStar parameters.}
\description{
Generic class for the LambdaStar parameters.
}
\section{Slots}{

\describe{
\item{\code{mu}}{The mean vector for the prior.}
}}

