% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-class.R
\name{names,bayesPO_fit-method}
\alias{names,bayesPO_fit-method}
\title{Names method for the bayesPO_fit class.}
\usage{
\S4method{names}{bayesPO_fit}(x)
}
\arguments{
\item{x}{A bayesPO_fit object.}
}
\description{
Names method for the bayesPO_fit class.
}
