% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-class.R
\name{names,bayesPO_initial-method}
\alias{names,bayesPO_initial-method}
\title{Names method for the bayesPO_initial class.}
\usage{
\S4method{names}{bayesPO_initial}(x)
}
\arguments{
\item{x}{The bayesPO_initial object.}
}
\description{
Names method for the bayesPO_initial class.
}
