% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\name{names,bayesPO_model-method}
\alias{names,bayesPO_model-method}
\title{Names method for the bayesPO_model class.}
\usage{
\S4method{names}{bayesPO_model}(x)
}
\arguments{
\item{x}{The bayesPO_model object.}
}
\description{
Names method for the bayesPO_model class.
}
