% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{names,bayesPO_prior-method}
\alias{names,bayesPO_prior-method}
\title{Names method for the bayesPO_prior class.}
\usage{
\S4method{names}{bayesPO_prior}(x)
}
\arguments{
\item{x}{The bayesPO_prior object.}
}
\description{
Names method for the bayesPO_prior class.
}
