% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{names,GammaPrior-method}
\alias{names,GammaPrior-method}
\title{Names method for the GammaPrior class.}
\usage{
\S4method{names}{GammaPrior}(x)
}
\arguments{
\item{x}{The GammaPrior object.}
}
\description{
Names method for the GammaPrior class.
}
