% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{names,NormalPrior-method}
\alias{names,NormalPrior-method}
\title{Names method for the NormalPrior class.}
\usage{
\S4method{names}{NormalPrior}(x)
}
\arguments{
\item{x}{The NormalPrior object.}
}
\description{
Names method for the NormalPrior class.
}
