% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{NormalPrior}
\alias{NormalPrior}
\title{Create a Normal prior object for model specification.}
\usage{
NormalPrior(mu, Sigma)
}
\arguments{
\item{mu}{The mean vector for the Normal distribution.}

\item{Sigma}{The covariance matrix for the Normal distribution.}
}
\value{
A \code{NormalPrior} object with adequate slots.
}
\description{
Constructor for \code{NormalPrior-class} objects
}
\details{
Matrix Sigma must be square and positive definite. Its dimensions
must match mu's length.
}
\examples{
NormalPrior(rep(0, 10), diag(10) * 10)
}
\seealso{
\code{\link{prior}}
}
