% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-class.R
\name{+,bayesPO_initial,ANY-method}
\alias{+,bayesPO_initial,ANY-method}
\title{Adding method for the bayesPO_initial class.}
\usage{
\S4method{+}{bayesPO_initial,ANY}(e1, e2)
}
\arguments{
\item{e1}{A bayesPO_initial object.}

\item{e2}{Another bayesPO_initial object.}
}
\value{
A list with the two objects.
}
\description{
Adding method for the bayesPO_initial class.
}
