% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-class.R
\name{+,bayesPO_initial,list-method}
\alias{+,bayesPO_initial,list-method}
\title{Adding method for the bayesPO_initial class with a list.}
\usage{
\S4method{+}{bayesPO_initial,list}(e1, e2)
}
\arguments{
\item{e1}{A bayesPO_initial object.}

\item{e2}{A list with bayesPO_initial objects.}
}
\value{
A list with the bayesPO_initial added.
}
\description{
Adding method for the bayesPO_initial class with a list.
}
