% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-class.R
\name{+,list,bayesPO_initial-method}
\alias{+,list,bayesPO_initial-method}
\title{Adding method for a list and a bayesPO_initial object.}
\usage{
\S4method{+}{list,bayesPO_initial}(e1, e2)
}
\arguments{
\item{e1}{A list with bayesPO_initial objects.}

\item{e2}{A bayesPO_initial object.}
}
\value{
A list with the bayesPO_initial added.
}
\description{
Adding method for a list and a bayesPO_initial object.
}
