% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-class.R
\name{print,bayesPO_fit-method}
\alias{print,bayesPO_fit-method}
\title{Print method}
\usage{
\S4method{print}{bayesPO_fit}(x, ...)
}
\arguments{
\item{x}{A bayesPO_fit object.}

\item{...}{Ignored.}
}
\description{
Print method
}
