% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-class.R
\name{print,bayesPO_initial-method}
\alias{print,bayesPO_initial-method}
\title{Print method for the bayesPO_initial class.}
\usage{
\S4method{print}{bayesPO_initial}(x, ...)
}
\arguments{
\item{x}{A bayesPO_initial object.}

\item{...}{Ignored.}
}
\description{
Print method for the bayesPO_initial class.
}
