% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{print,bayesPO_prior-method}
\alias{print,bayesPO_prior-method}
\title{Print method for the bayesPO_prior class.}
\usage{
\S4method{print}{bayesPO_prior}(x, ...)
}
\arguments{
\item{x}{The bayesPO_prior object.}

\item{...}{Ignored.}
}
\description{
Print method for the bayesPO_prior class.
}
