% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{print,BetaDeltaPrior-method}
\alias{print,BetaDeltaPrior-method}
\title{Print method for the BetaDeltaPrior class.}
\usage{
\S4method{print}{BetaDeltaPrior}(x, ...)
}
\arguments{
\item{x}{The BetaDeltaPrior object.}

\item{...}{Ignored.}
}
\description{
Print method for the BetaDeltaPrior class.
}
