% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{print,GammaPrior-method}
\alias{print,GammaPrior-method}
\title{Print method for the GammaPrior class.}
\usage{
\S4method{print}{GammaPrior}(x, ...)
}
\arguments{
\item{x}{The GammaPrior object.}

\item{...}{Ignored.}
}
\description{
Print method for the GammaPrior class.
}
