% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{print,NormalPrior-method}
\alias{print,NormalPrior-method}
\title{Print method for the NormalPrior class.}
\usage{
\S4method{print}{NormalPrior}(x, ...)
}
\arguments{
\item{x}{The NormalPrior object.}

\item{...}{Ignored.}
}
\description{
Print method for the NormalPrior class.
}
