% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{retrievePars,GammaPrior-method}
\alias{retrievePars,GammaPrior-method}
\title{Retrieve method for the GammaPrior class.}
\usage{
\S4method{retrievePars}{GammaPrior}(object)
}
\arguments{
\item{object}{The GammaPrior object.}
}
\description{
Retrieve method for the GammaPrior class.
}
