% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{retrievePars,NormalPrior-method}
\alias{retrievePars,NormalPrior-method}
\title{Retrieve method for the NormalPrior class.}
\usage{
\S4method{retrievePars}{NormalPrior}(object)
}
\arguments{
\item{object}{The NormalPrior object.}
}
\description{
Retrieve method for the NormalPrior class.
}
