% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{retrievePars}
\alias{retrievePars}
\title{Retrieve the parameters in an adequate format for the C++ code.}
\usage{
retrievePars(object)
}
\arguments{
\item{object}{The object from which the parameters will be extracted.}
}
\description{
Retrieve the parameters in an adequate format for the C++ code.
}
