% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-class.R
\name{show,bayesPO_fit-method}
\alias{show,bayesPO_fit-method}
\title{Show method for the bayesPO_fit class.}
\usage{
\S4method{show}{bayesPO_fit}(object)
}
\arguments{
\item{object}{A bayesPO_fit object.}
}
\description{
Show method for the bayesPO_fit class.
}
