% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-class.R
\name{show,bayesPO_initial-method}
\alias{show,bayesPO_initial-method}
\title{Show method for the bayesPO_initial class.}
\usage{
\S4method{show}{bayesPO_initial}(object)
}
\arguments{
\item{object}{A bayesPO_initial object.}
}
\description{
Show method for the bayesPO_initial class.
}
