% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{show,bayesPO_prior-method}
\alias{show,bayesPO_prior-method}
\title{Show method for the bayesPO_prior class.}
\usage{
\S4method{show}{bayesPO_prior}(object)
}
\arguments{
\item{object}{The bayesPO_prior object.}
}
\description{
Show method for the bayesPO_prior class.
}
