% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{show,BetaDeltaPrior-method}
\alias{show,BetaDeltaPrior-method}
\title{Show method for the BetaDeltaPrior class.}
\usage{
\S4method{show}{BetaDeltaPrior}(object)
}
\arguments{
\item{object}{the BetaDeltaPrior object.}
}
\description{
Show method for the BetaDeltaPrior class.
}
