% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{show,GammaPrior-method}
\alias{show,GammaPrior-method}
\title{Show method for the GammaPrior class.}
\usage{
\S4method{show}{GammaPrior}(object)
}
\arguments{
\item{object}{The GammaPrior object.}
}
\description{
Show method for the GammaPrior class.
}
