% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{show,NormalPrior-method}
\alias{show,NormalPrior-method}
\title{Show methodfor the NormalPrior class.}
\usage{
\S4method{show}{NormalPrior}(object)
}
\arguments{
\item{object}{The NormalPrior object.}
}
\description{
Show methodfor the NormalPrior class.
}
