% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-class.R
\name{[[,bayesPO_fit-method}
\alias{[[,bayesPO_fit-method}
\title{The '[[' method for the bayesPO_fit class.}
\usage{
\S4method{[[}{bayesPO_fit}(x, i)
}
\arguments{
\item{x}{A bayesPO_fit object.}

\item{i}{The requested slot.}
}
\description{
The '[[' method for the bayesPO_fit class.
}
