% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-class.R
\name{summary,bayesPO_fit-method}
\alias{summary,bayesPO_fit-method}
\title{Summary method for the bayesPO_fit class.}
\usage{
\S4method{summary}{bayesPO_fit}(object, ...)
}
\arguments{
\item{object}{A bayesPO_fit object.}

\item{...}{Ignored.}
}
\value{
A matrix with the summary.
}
\description{
Summary method for the bayesPO_fit class.
}
