% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-class.R
\name{*,bayesPO_initial,numeric-method}
\alias{*,bayesPO_initial,numeric-method}
\title{Multiplying methods for the bayesPO_initial class.}
\usage{
\S4method{*}{bayesPO_initial,numeric}(e1, e2)
}
\arguments{
\item{e1}{A bayesPO_initial object.}

\item{e2}{A positive integer.}
}
\value{
A list with \code{e2} random initial values.
}
\description{
If the initial values were random, then a list is created with the requested
amount of random initial values.
}
