% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-class.R
\name{*,numeric,bayesPO_initial-method}
\alias{*,numeric,bayesPO_initial-method}
\title{Multiplying methods for the bayesPO_initial class.}
\usage{
\S4method{*}{numeric,bayesPO_initial}(e1, e2)
}
\arguments{
\item{e1}{A positive integer.}

\item{e2}{A bayesPO_initial object.}
}
\value{
A list with \code{e1} random initial values.
}
\description{
If the initial values were random, then a list is created with the requested
amount of random initial values.
}
