% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-class.R
\docType{class}
\name{bayesPO_initial-class}
\alias{bayesPO_initial-class}
\title{Class for the initial values for the MCMC for the bayesPO package}
\description{
Class for the initial values for the MCMC for the bayesPO package
}
\section{Slots}{

\describe{
\item{\code{beta}}{Initial values for beta.}

\item{\code{delta}}{Initial values for delta.}

\item{\code{lambdaStar}}{Initial values for lambdaStar.}

\item{\code{tag}}{Indicates the source of the initial values.}
}}

