% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-class.R
\docType{class}
\name{bayesPO_initial-class}
\alias{bayesPO_initial-class}
\alias{names,bayesPO_initial-method}
\alias{$,bayesPO_initial-method}
\alias{+,bayesPO_initial,ANY-method}
\alias{+,bayesPO_initial,list-method}
\alias{+,list,bayesPO_initial-method}
\alias{*,bayesPO_initial,numeric-method}
\alias{*,numeric,bayesPO_initial-method}
\alias{show,bayesPO_initial-method}
\alias{print,bayesPO_initial-method}
\alias{print.bayesPO_initial}
\title{Class for the initial values for the MCMC for the bayesPO package}
\usage{
\S4method{names}{bayesPO_initial}(x)

\S4method{$}{bayesPO_initial}(x, name)

\S4method{+}{bayesPO_initial,ANY}(e1, e2)

\S4method{+}{bayesPO_initial,list}(e1, e2)

\S4method{+}{list,bayesPO_initial}(e1, e2)

\S4method{*}{bayesPO_initial,numeric}(e1, e2)

\S4method{*}{numeric,bayesPO_initial}(e1, e2)

\S4method{show}{bayesPO_initial}(object)

\S4method{print}{bayesPO_initial}(x, ...)

\method{print}{bayesPO_initial}(x, ...)
}
\arguments{
\item{x}{The bayesPO_initial object.}

\item{name}{The requested slot.}

\item{e1}{A bayesPO_initial object.}

\item{e2}{Another bayesPO_initial object or a list with bayesPO_initial
objects for \strong{+} and a positive integer for \strong{*}. e1 and e2
can be switched (+ and * are commutative).}

\item{object}{A bayesPO_initial object.}

\item{...}{Currently unused.}
}
\value{
\strong{\code{names}}: A character vector with the initialized
parameter names.

\strong{\code{`$`}}: The requested initial value (in case of
LambdaStar) or values (in case of Beta or Delta).

\strong{\code{+}}: A list with the objects. Useful to start the
\code{fit_bayesPO} function, as it requires a list of initial values.

\strong{\code{*}}: A list with \code{e2} random initial values.

\strong{\code{show}} and \strong{\code{print}}: The invisible object.
}
\description{
Class for the initial values for the MCMC for the bayesPO package
}
\section{Fields}{

\describe{
\item{\code{beta}}{Initial values for beta.}

\item{\code{delta}}{Initial values for delta.}

\item{\code{lambdaStar}}{Initial values for lambdaStar.}

\item{\code{tag}}{Indicates the source of the initial values.}
}}

