% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\docType{class}
\name{bayesPO_prior-class}
\alias{bayesPO_prior-class}
\alias{names,bayesPO_prior-method}
\alias{$,bayesPO_prior-method}
\alias{show,bayesPO_prior-method}
\alias{print,bayesPO_prior-method}
\alias{print.bayesPO_prior}
\alias{$<-,bayesPO_prior-method}
\title{Joint prior class for the bayesPO package parameters}
\usage{
\S4method{names}{bayesPO_prior}(x)

\S4method{$}{bayesPO_prior}(x, name)

\S4method{show}{bayesPO_prior}(object)

\S4method{print}{bayesPO_prior}(x, ...)

\method{print}{bayesPO_prior}(x, ...)

\S4method{$}{bayesPO_prior}(x, name)

\S4method{$}{bayesPO_prior}(x, name) <- value
}
\arguments{
\item{x}{The bayesPO_prior object.}

\item{name}{The requested slot.}

\item{object}{The bayesPO_prior object.}

\item{...}{Ignored.}

\item{value}{New value.}
}
\value{
\strong{\code{names}}: A character vector with the model parameters
names.

\strong{\code{`$`}}: The requested slot's value.

\strong{\code{`$<-`}}: The new object with the updated slot.
}
\description{
Objects of this class are the joining of independent priors for Beta, Delta
and LambdaStar. They can be used in the \code{fit_bayesPO} function.
}
\section{Fields}{

\describe{
\item{\code{beta}}{An object of a class which inherits the \code{BetaDeltaPrior} S4
class with the appropriate Beta prior.}

\item{\code{delta}}{An object of a class which inherits the \code{BetaDeltaPrior} S4
class with the appropriate Delta prior.}

\item{\code{lambdaStar}}{An object of a class which inherits the
\code{LambdaStarPrior} S4 class with the appropriate LambdaStar prior.}
}}

