% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\docType{class}
\name{BetaDeltaPrior-class}
\alias{BetaDeltaPrior-class}
\alias{show,BetaDeltaPrior-method}
\alias{print,BetaDeltaPrior-method}
\alias{print.BetaDeltaPrior}
\title{Generic class for the beta and delta parameters.}
\usage{
\S4method{show}{BetaDeltaPrior}(object)

\S4method{print}{BetaDeltaPrior}(x, ...)

\method{print}{BetaDeltaPrior}(x, ...)
}
\arguments{
\item{object}{The BetaDeltaPrior object.}

\item{x}{The BetaDeltaPrior object.}

\item{...}{Ignored.}
}
\value{
\strong{\code{show}} and \strong{\code{print}}: The invisible object.
}
\description{
Generic class for the beta and delta parameters.
}
\section{Fields}{

\describe{
\item{\code{family}}{The family of distributions of the prior.}
}}

