% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesPO.R
\name{fit_bayesPO}
\alias{fit_bayesPO}
\title{Fit Presence-Only data using a Bayesian Poisson Process model}
\usage{
fit_bayesPO(object, background, mcmc_setup = list(iter = 5000), ...)

\S4method{fit_bayesPO}{bayesPO_model,matrix}(object, background, mcmc_setup, area = 1, cores = 1)
}
\arguments{
\item{object}{A bayesPO_model object.}

\item{background}{A matrix with the covariates values in the region background.}

\item{mcmc_setup}{A list with the components \code{burnin}, \code{thin} and
\code{iter} where only the latter is mandatory.}

\item{area}{A positive number with the region's area}

\item{data}{A list containing a matrix with the covariates for the observed
points, a matrix with the covariates for the background, the region's area
and the column choices for the intensity and observability sets. See details.}

\item{intensityLink}{A string with the chosen link function for the
intensity process. Currently only \code{"logit"} is accepted.}

\item{intensityPrior}{A string with the chosen prior distribution for the
intensity effects \code{beta}. Currently only \code{"normal"} is accepted.}

\item{observabilityLink}{A string with the chosen link function for the
observability process. Currently only \code{"logit"} is accepted.}

\item{observabilityPrior}{A string with the chosen prior distribution for the
observability effects \code{delta}. Currently only \code{"normal"} is accepted.}

\item{init}{The initial values for the chains. The string \code{"random"} can
be used to use randomly chosen values. See details for instructions on how
to provide specific values.}

\item{prior}{Parameters for the prior. It must be a \code{list} containing
a list for each of the parameters \code{beta}, \code{delta} and
\code{lambdaStar}. See details for more information.}

\item{chains}{Number of independent chains to run. If more than 1 are
requested and argument \code{init} is not \code{"random"}, initial values
must be provided for all chains. See details for more instructions about
this.}
}
\value{
An object of class \code{"bayesPO_fit"}. It has its own print and
summary methods. It has methods for as.array, as.matrix and as.data.frame,
thus can be used with the bayesplot package (as.array) or ggplot2
(as.data.frame).
}
\description{
The model uses a data augmentation scheme to avoid performing approximations
on the likelihood function.
}
\details{
## Input instructions
A few notes about properly feeding information to the model are provided
here.
### data
The data input must be a list with the following elements:
\itemize{
\item{observed}{A matrix containing the covariate values for the
presence-only observed points.}
\item{background}{A matrix containing the covariate values for the
background. The values must span over the entire studied region.}
\item{area}{A single positive value informing the studied region's area.
Changing this value will not (greatly) affect the estimation of \code{beta}
and \code{delta}, but it will affect \code{lambdaStar} and predictions using
the model. This means that a correct value must be provided for adequate
predictions.}
\item{intensity}{A vector indicating which columns explain the species'
intensity, that is, its occurrence. It can be either a numeric vector or
a character vector. If numeric, it indicates which column positions are
to be included as intensity covariates. Arguments \code{observed} and
\code{background} must have identical columns positions for the covariates.
If character, it indicates the names of the columns to be included as
intensity covariates. Both \code{observed} and \code{background} must have
the indicated columns, but they do not need to be in the same position.
Parameters \code{beta} measure their effects in the model.}
\item{observability}{A vector indicating which columns explain the species'
observability, that is, the probability that its occurrences are observed.
It can be either a numeric vector or a character vector. If numeric, it
indicates which column positions are to be included as observability
covariates. Arguments \code{observed} and \code{background} must have
identical columns positions for the covariates. If character, it indicates
the names of the columns to be included as observability covariates.
Both \code{observed} and \code{background} must have the indicated columns,
but they do not need to be in the same position. Parameters \code{delta}
measure their effects in the model.}
}

### init
The initial values for the MCMC chains can be chosen to be random with the
\code{"random"} string. However, if the user wishes to input specific values,
they must be included in a list inside a list. The outer list is to group
up the initial values for the different chains.

The inner ones inform the model the initial values for each chain, so the
first list includes the values for the first chain, and so forth. If any of
the inner lists does not contain one of them, a random value is included.

For \code{beta} and \code{delta}, a single value can be given, at which case
the same value is used for the entire vector. Otherwise, the size of
\code{beta} and \code{delta} must match the selections of one plus intensity
and observability covariates, respectively. The extra value is meant for the
intercepts, which go first in the order, meaning that the first initial
value of each vector is associated with the intercepts.

For \code{lambdaStar} only a single value per chain must be provided.

### prior
Different prior choices imply different parameter choices.

For the \code{beta} and \code{delta} "normal" choice, a mean vector and
covariance matrix must be provided. The \code{beta} mean vector must be one
plus the choice size of intensity covariates and the \code{delta} mean
vector must be one plus the choice size of observability covariates. The
extra one is for the intercept, which comes first, meaning that the first
value of the vector is the prior mean of the intercept. The covariance
matrices must be symmetric and positive definite.

Alternatively, a single value is provided. For the mean, this means that
the prior mean is equal to the chosen value for all \code{beta} parameters.
The logic is analogous for \code{delta.} If the covariance is a single
positive number, then the respective matrix is an identity multiplied by
this number.

For \code{lambdaStar} the prior is a gamma distribution with parameters a and b
for the shape and rate parameters, respectively.

### mcmc_setup
}
