% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial-class.R
\name{initialize,bayesPO_initial-method}
\alias{initialize,bayesPO_initial-method}
\title{Initialize method for the bayesPO_initial class.}
\usage{
\S4method{initialize}{bayesPO_initial}(.Object, beta, delta, lambdaStar, tag)
}
\arguments{
\item{.Object}{The bayesPO_initial object.}

\item{beta}{A vector with the initial values for beta.}

\item{delta}{A vector with the initial values for delta.}

\item{lambdaStar}{A positive with the initial value for lambdaStar.}

\item{tag}{A string indicating the origin for the initial values.}
}
\description{
Initialize method for the bayesPO_initial class.
}
