% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-class.R
\name{initialize,bayesPO_model-method}
\alias{initialize,bayesPO_model-method}
\title{Initialize method for the bayesPO_model class.}
\usage{
\S4method{initialize}{bayesPO_model}(
  .Object,
  po,
  intensityLink,
  intensitySelection,
  observabilityLink,
  observabilitySelection,
  init,
  prior,
  iSelectedColumns,
  oSelectedColumns
)
}
\arguments{
\item{.Object}{The bayesPO_model object.}

\item{po}{A matrix containing the covariates observed in the presence-only locations.}

\item{intensityLink}{The link for the intensity component of the model.}

\item{intensitySelection}{Column indexes in the \code{po} matrix for the intensity covariates.}

\item{observabilityLink}{The link for the observability component of the model.}

\item{observabilitySelection}{Column indexes in the \code{po} matrix for the observability covariates.}

\item{init}{A list with initial values.}

\item{prior}{An object with the joint prior.}

\item{iSelectedColumn}{The names of the intensity columns selected.}

\item{oSelectedColumn}{The names of the observability columns selected.}
}
\description{
Initialize method for the bayesPO_model class.
}
