% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\docType{class}
\name{NormalPrior-class}
\alias{NormalPrior-class}
\alias{print.NormalPrior}
\title{Normal prior class for Beta and Delta parameters.}
\usage{
\S4method{names}{NormalPrior}(x)

\S4method{$}{NormalPrior}(x, name)

\S4method{$}{NormalPrior}(x, name) <- value

\S4method{show}{NormalPrior}(object)

\S4method{print}{NormalPrior}(x, ...)

print.NormalPrior(x, ...)
}
\arguments{
\item{x}{The NormalPrior object.}

\item{name}{The slot to be changed.}

\item{value}{New value.}

\item{object}{The NormalPrior object.}

\item{...}{Ignored.}
}
\description{
This is used to represent the prior for Beta and Delta individually. They
still need to be joined to be used in a model.
}
\section{Slots}{

\describe{
\item{\code{mu}}{The mean vector for the prior.}

\item{\code{Sigma}}{The covariance matrix for the prior.}
}}

