% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\docType{class}
\name{NormalPrior-class}
\alias{NormalPrior-class}
\title{Normal prior class for Beta and Delta parameters.}
\description{
Normal prior class for Beta and Delta parameters.
}
\section{Slots}{

\describe{
\item{\code{mu}}{The mean vector for the prior.}

\item{\code{Sigma}}{The covariance matrix for the prior.}
}}

