% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior-class.R
\name{prior}
\alias{prior}
\title{Build a joint prior for bayesPO model parameters}
\usage{
prior(beta, delta, lambdaStar)
}
\arguments{
\item{beta}{An S4 object whose class inherits from \code{BetaDeltaPrior}.}

\item{delta}{An S4 object whose class inherits from \code{BetaDeltaPrior}.}

\item{lambdaStar}{An S4 object whose class inherits from \code{LambdaStarPrior}.}
}
\description{
Constructor for \code{bayesPO_prior} objects, which is used in the
\code{bayesPO_fit} function. The generated prior is so that Beta, Delta
and LambdaStar are indepdendent a priori.
}
\seealso{
\code{\link{fit_bayesPO}}, \code{\link{NormalPrior}} and
\code{\link{GammaPrior}}
}
